Fix up & trim down dependencies in composer.json
Move wikipeg from a dev dependency to a 'real' dependency, because the
wikipeg runtime is needed to run the parser.  Removed a bunch of
unnecessary libraries dragged in from the skeleton I used, and update
the version numbers on others to match the most recent versions being
used in Wikimedia production.

Change-Id: Iade4d7aca8cd3a5e388e5cf43c05bd284b0f4455